Refactor imports and enhance weight standardization function in neural network module
